🚚move: q and a file (#920)
